+++ scout/server/blueprints/institutes/views.py
@@ -218,6 +218,12 @@ def gene_variants(institute_id):
                 "Gene not included in clinical list: {}".format(", ".join(non_clinical_symbols)),
                 "warning",
             )
+
+        if hgnc_symbols == []:
+            # If there are not genes to search, return to previous page with a warning
+            flash("No valid gene provided for variant search.", "warning")
+            return redirect(request.referrer)
+
         form.hgnc_symbols.data = hgnc_symbols
 
         LOG.debug("query {}".format(form.data))